# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: VNC Remote Secure API
  description: |
    REST API for VNC Remote Secure — a self-hosted remote access platform
    with browser-based VNC, Web Terminal, and health monitoring.

    ## Authentication
    Health, metrics, and audit endpoints accept a Bearer token
    (`HEALTH_AUTH_TOKEN`). When the token is unset, access is open only
    while every health-serving bind (`HEALTH_WEB_HOST`, `USER_UI_HOST`,
    `BIND_HOST`) is loopback — a public bind without a token fails
    closed with 401. The Flask user-management UI
    uses a `vnc_session` cookie issued by `/login` (with CSRF
    protection); that scheme is not covered by this spec and is not
    accepted by the health server. MFA (TOTP) is required for
    user-management when `MFA_REQUIRED=true`.

    ## Security
    - Backends bind to 127.0.0.1 only
    - Public access goes through nginx reverse proxy
    - All responses include security headers (HSTS, CSP, X-Frame-Options)
  version: 0.2.0
  contact:
    url: https://github.com/alesanfe/vnc-remote-secure/issues
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://127.0.0.1:8080
    description: Linux (health service)
  - url: http://127.0.0.1:8090
    description: Windows (health service)
  - url: http://127.0.0.1:8081
    description: User-management UI (metrics and audit endpoints)
  - url: https://{domain}
    description: Public (via nginx reverse proxy)
    variables:
      domain:
...
paths:
...
  /audit:
    get:
      tags: [Audit]
      summary: Recent audit log entries
      description: Returns recent structured audit log entries.
      security:
        - bearerToken: []
      parameters:
        - name: limit
          in: query
          description: Maximum number of entries to return
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 1000
        - name: event
          in: query
          description: Filter by event type
          schema:
            type: string
      responses:
        '200':
          description: Audit log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuditEntry'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /audit/verify:
    get:
      tags: [Audit]
      summary: Verify audit log chain integrity
      description: Verifies the tamper-evident hash chain of the audit log.
      security:
        - bearerToken: []
      responses:
        '200':
          description: Chain verification result
          content:
            application/json:
              schema:
                type: object
                properties:
                  intact:
                    type: boolean
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
components:
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Bearer token matching the `HEALTH_AUTH_TOKEN` environment variable.
        When the token is unset, the health server accepts unauthenticated
        requests (intended for localhost-only binding).
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: boolean
          example: true
        code:
          type: string
          example: ERR_AUTH_REQUIRED
        message:
          type: string
          example: Unauthorized
...
    AuditEntry:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        event:
          type: string
          example: login
        user:
          type: string
        ip:
          type: string
        result:
          type: string
          enum: [success, failure]
        detail:
          type: string
        hash:
          type: string
          description: SHA-256 chain hash for tamper-evidence